requirements.txt
polib==1.1.0
google-cloud-translate==1.3.1
